# chore(validation): improve field merging perf